enhance(stories migrate): Fix storybook Parameters type augmentation to remove tsconfig typeRoots override
