+++ stories/index.d.ts
@@ -1,0 +1,1 @@
+import '../.storybook/mock'
