+++ tsconfig.json
@@ -21,10 +21,6 @@
         "src/*"
       ]
     },
-    "typeRoots": [
-      "./node_modules/@types",
-      "./.storybook/mock",
-    ]
   },
   "include": [
     "src/**/*.d.ts",
